Rename files and String constants for Status implemented
